requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6


motor==3.3.2
pymongo==4.6.1

pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0


pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

google-generativeai==0.3.2
pillow==10.2.0

requests==2.31.0
httpx==0.26.0

websockets==12.0


python-magic==0.4.27
aiofiles==23.2.1


python-dateutil==2.8.2


python-dotenv==1.0.0


starlette==0.35.1


pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0